[Card scanning, opencv.js] Card scanning using opencv.js filtering
The process of scanning:

1. Detect the edges of the credit card.
2. Localize the four groupings of four digits,
pertaining to the sixteen digits on the credit card.
3. Apply OCR to recognize the sixteen digits on the credit card.

Note: There should be a strong contrast between the numbers
and the card background otherwise digits can not be detected.
Also, to detect edge of the card, the video background should
have another color than the card color.

Python tutorial of card scanning with opencv:
https://www.pyimagesearch.com/2017/07/17/credit-card-ocr-with-opencv-and-python/